Fixed bad merge and things

+++ view/renderer/renderer.cpp
@@ -1,25 +1,27 @@
 #include "renderer.h"
 
-Renderer::Renderer(){}
+Renderer::Renderer() {
+}
 
-QPixmap Renderer::renderGameObject(const QSharedPointer<GameObject>& object) {
+QPixmap
+Renderer::renderGameObject(const QSharedPointer<GameObject>& object) {
 
-    for (auto data: object->getAllData()){
+    for (auto data : object->getAllData()) {
         switch (data[GameObject::DataRole::Type].toInt()) {
-        case GameObject::ObjectType::Tile:
+        case static_cast<int>(GameObject::ObjectType::Tile):
             return renderTile(data);
-        case GameObject::ObjectType::Doorway:
+        case static_cast<int>(GameObject::ObjectType::Doorway):
             return renderDoorway(data);
-        case GameObject::ObjectType::HealthPack:
+        case static_cast<int>(GameObject::ObjectType::HealthPack):
             return renderHealthPack(data);
-        case GameObject::ObjectType::Protagonist:
+        case static_cast<int>(GameObject::ObjectType::Protagonist):
             return renderProtagonist(data);
-//        case GameObject::ObjectType::Enemy:
-//            return renderEnemy(data);
-//        case GameObject::ObjectType::PoisonEnemy:
-//            return renderEnemy(data);
-//        case GameObject::ObjectType::MovingEnemy:
-//            return renderMovingEnemy(data);
+            //        case GameObject::ObjectType::Enemy:
+            //            return renderEnemy(data);
+            //        case GameObject::ObjectType::PoisonEnemy:
+            //            return renderEnemy(data);
+            //        case GameObject::ObjectType::MovingEnemy:
+            //            return renderMovingEnemy(data);
         default:
             // Handle default case or unknown types
             return renderEnemy(data);
